README.md
Describes how to play all the games in this repository. It runs from a browser. Use Google Chrome for better performance.

Fire up http://www.codeskulptor.org/ on Chrome browser and use the code name along with it.
Eg: If you want to execute the game pong_game.py, execute it as mentioned below -
http://www.codeskulptor.org/#pong_game.py